+++ Modules/Intent.Modules.Common.Java/Templates/JavaTemplateBaseExtensionMethods.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Intent.Code.Weaving.Java.Editor;
 using Intent.Engine;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.Types.Api;
@@ -36,27 +35,6 @@ namespace Intent.Modules.Common.Java.Templates
         public static string GetPackageFolderPath<TModel>(this JavaTemplateBase<TModel> template, params string[] additionalFolders)
         {
             return template.GetFolderPath(additionalFolders);
-        }
-
-        /// <summary>
-        /// Determines all imports for the provided <paramref name="template"/> and adds them to
-        /// the provided <paramref name="file"/>.
-        /// </summary>
-        public static void ResolveAndAddImports<TModel>(this JavaTemplateBase<TModel> template, JavaFile file)
-        {
-            var imports = template.GetAll<IDeclareImports, string>(item => item.DeclareImports())
-                .Where(import => !string.IsNullOrWhiteSpace(import))
-                .Distinct();
-
-            foreach (var import in imports)
-            {
-                if (file.ImportExists(import) || template.Package == import.RemoveSuffix("." + import.Split('.').Last()))
-                {
-                    continue;
-                }
-
-                file.AddImport($"{Environment.NewLine}import {import};");
-            }
         }
 
         /// <summary>